Separate db functions into specific files
