+++ src/db/tagCloud.ts
@@ -1,0 +1,7 @@
+import { Client } from "pg";
+
+export async function getTagCloud(client: Client) {
+    const result = await client.query(`SELECT tag_name FROM tags_cloud;`);
+
+    return result;
+}
